requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2024.1
Django==5.0.3
django-crispy-forms==2.1
djangorestframework==3.15.1
pillow==10.2.0
sqlparse==0.4.4
typing_extensions==4.10.0
